Fix script due to Youtube API Auth changes.
Update to make all accesses to Youtube V3 API use authentication
as the script stopped working since it used to use a simple GET
request with an API key to access the initial snippet fetching logic.

Now we use the authenticated API logic for fetching both the initial
video information as well as the tags.