Fixed session handling to work better with Wordpress
And added example code for making the video player responsive.

While using this plugin, the Wordpress Health screen showed that
a session was being left open, which interfered with the REST API
and loopback calls.
By closing the PHP session after we are done with it, this seems to
fix the issue.

Also, added example code that can be enabled to make the video
player responsive. Any existing blog posts will need to be updated
manually (such as by using a Search and Replace plugin) if the new
video player code is enabled, as this plugin only will affect future videos
that are added to your blog.